+++ easy/0605_can_place_flowers/solution.cpp
@@ -3,9 +3,9 @@
 class Solution {
 public:
   bool canPlaceFlowers(std::vector<int> &flowerbed, int n) {
-    const int flowerbedLen = flowerbed.size();
+    const size_t flowerbedLen = flowerbed.size();
 
-    for (int i = 0; n && i < flowerbedLen; ++i) {
+    for (size_t i = 0; n && i < flowerbedLen; ++i) {
       if (flowerbed[i] == 0 && (i - 1 == -1 || flowerbed[i - 1] == 0) &&
           (i + 1 == flowerbedLen || flowerbed[i + 1] == 0)) {
         flowerbed[i++] = 1;
