+++ easy/1984_minimum_difference_between_highest_and_lowest_of_k_scores/solution.cpp
@@ -7,7 +7,7 @@ public:
 
     int minDiff = 100000;
 
-    for (int i = 0; i < nums.size() - k + 1; ++i)
+    for (size_t i = 0; i < nums.size() - k + 1; ++i)
       minDiff = std::min(minDiff, nums[i + k - 1] - nums[i]);
 
     return minDiff;
